Create Jump_To_The_End(dp).cpp
You are given a sequence of nonnegative integers 
A
1
,
A
2
,
…
,
A
N
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 .

There is a pawn placed on the first element of the sequence, i.e. at position 
1
1. When the pawn is on the position numbered with 
x
x, it can jump at most 
A
x
A 
x
​
  places to the right. In other words, if 
A
x
>
0
A 
x
​
 >0, the pawn can jump to any position from 
x
+
1
x+1 to 
x
+
A
x
x+A 
x
​
 . If 
A
x
=
0
A 
x
​
 =0, the pawn can not move from the position numbered with 
x
x at all.

Find the maximum possible number of jumps using which it is possible to place the pawn on the last position, i.e. the position numbered with 
N
N, or determine that it is impossible to reach it.

Input Format
The first line of the input contains a single integer 
T
T - the number of test cases. The description of 
T
T test cases follows.

The first line of each test case contains an integer 
N
N.

The second line contains 
N
N space-separated integers 
A
1
,
A
2
,
…
,
A
N
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
 .

Output Format
For each test case, print a single line containing one integer. In case it is possible to reach the last position, this integer should be the maximum possible number of jumps in which the pawn can be placed on the last position. Otherwise, this integer should be 
−
1
−1.