Create StackSort.cpp
You have an array 
A
A of integers of size 
N
N, an array 
B
B ( initially empty ) and a stack 
S
S ( initially empty ). You are allowed to do the following operations :

Take the first element of array 
A
A and push it into 
S
S and remove it from 
A
A.
Take the top element from stack 
S
S, append it to the end of array 
B
B and remove it from 
S
S.
You have to tell if it possible to move all the elements of array 
A
A to array 
B
B using the above operations such that finally the array 
B
B is sorted in ascending order.

Input Format
First line will contain 
T
T, number of testcases. Then the testcases follow.
First line of each testcase contains a single integer 
N
N.
Second line of each testcase contains 
N
N distinct integers : 
A
1
,
A
2
.
.
.
A
N
A 
1
​
 ,A 
2
​
 ...A 
N
​
 .
Output Format
For each testcase, if it possible to move all the elements of array 
A
A to array 
B
B using the above operations such that finally, the array 
B
B is sorted in ascending order, print "YES" (without quotes), else print "NO" (without quotes).