Create Find_the_Pattern.cpp
You are given two strings 
S
S and 
P
P both consisting of lowercase letters of the English alphabet. Find if the string 
P
P occurs as a substring in the string 
S
S.

A substring is a contiguous sequence of characters within a string.

Input Format
The first line of the input contains a single integer 
T
T - the number of test cases. The description of 
T
T test cases follows.

The first line of each test case contains a single string 
S
S.

The second line of each test case contains a single string 
P
P.

Output Format
For each test case, print a single line containing one integer. That integer should be 
1
1 if the string 
P
P occurs as a substring in the string 
S
S and 
0
0 otherwise.
Constraints
1
≤
T
≤
100
1≤T≤100
1
≤
∣
P
∣
≤
∣
S
∣
≤
2
⋅
1
0
5
1≤∣P∣≤∣S∣≤2⋅10 
5
 
the characters in 
S
S and 
P
P are lowercase letters of the English alphabet
the sum of 
∣
S
∣
∣S∣ over all test cases does not exceed 
4
⋅
1
0
5
4⋅10 
5