Create Minimum-Circular-Jumps-to-Escape(I).cpp
A group of N people are seated around a circular 
table to play a game.
 The game involves jumping from one chair to 
another. Each person sitting on chair i can jump A[i] 
chairs to either the right or left in one jump where 0 
< i < N+1.
 Bob, sitting on chair X, needs to reach chair Y, where 
the escape door is located.
 Find the minimum number of jumps required to 
reach chair Y from chair X. If this is impossible using 
the given jump distances, then return -1.
 Input Format
 The first line contains an integer, N, denoting the 
number of people playing the game.
 The next line contains an integer, X, denoting the 
chair on which Bob is seated.
 The next line contains an integer, Y, denoting the 
chair which Bob wants to reach.
 Each line i of the N subsequent lines (where 1 ≤ i ≤ 
N) contains an integer describing A[i] is the number 
of chairs the person sitting in chair number i can 
jump either right or left.