Create Digit-to-Character-Encoding-with-Vowel-Constraints(A).cpp
Ratan has invented a special encoding system. He wants to encode a positive integer n into a string using the following rules:

Prime digits (2, 3, 5, 7) must be mapped to unique vowels from the set {a, e, i, o, u}.

Non-prime digits (0, 1, 4, 6, 8, 9) must be mapped to consonants, skipping all vowels.

Each digit should be mapped to a unique character, and repeated digits must be mapped consistently.

The digits of the number are encoded from right to left, and the resulting string should be reversed at the end to restore the original order.

If there are not enough vowels or consonants left for unique mapping, you can assume the input will always have a valid solution within the character set 'a' to 'z'.